apps/api/requirements.txt
uvicorn
fastapi
websockets
google-genai
httpx
python-dotenv
